+++ custom_components/hacs/repositories/helpers/methods/__init__.py
@@ -1,9 +1,11 @@
 # pylint: disable=missing-class-docstring,missing-module-docstring,missing-function-docstring,no-member
 from .exsist_on_local_fs import RepositoryMethodExsistOnLocalFS
 from .reinstall_if_needed import RepositoryMethodReinstallIfNeeded
-from .install import RepositoryMethodInstall
-from .pre_install import RepositoryMethodPreInstall
-from .post_install import RepositoryMethodPostInstall
+from .installation import (
+    RepositoryMethodPreInstall,
+    RepositoryMethodInstall,
+    RepositoryMethodPostInstall,
+)
 
 
 class RepositoryHelperMethods(
