+++ custom_components/hacs/repositories/plugin.py
@@ -46,18 +46,6 @@ class HacsPlugin(HacsRepository):
                     self.logger.error(error)
         return self.validate.success
 
-    async def registration(self, ref=None):
-        """Registration."""
-        if ref is not None:
-            self.data.selected_tag = ref
-            self.ref = ref
-            self.force_branch = True
-        if not await self.validate_repository():
-            return False
-
-        # Run common registration steps.
-        await self.common_registration()
-
     async def update_repository(self, ignore_issues=False):
         """Update."""
         await self.common_update(ignore_issues)
